fullstack_app/backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0.post1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
openai==1.3.5
python-docx==0.8.11
pdfplumber==0.9.0
pandas==2.1.4
python-magic==0.4.27
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==1.10.12